Module dusk_hades::strategies

Strategies implemented for the Hades252 algorithm. This module contains an implementation of the Hades252 strategy algorithm specifically designed to work outside of Rank 1 Constraint Systems (R1CS) or other custom Constraint Systems such as Add/Mul/Custom plonk gate-circuits.

The inputs of the permutation function have to be explicitly over the BlsScalar Field of the bls12_381 curve so working over Fq = 0x73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001.

Re-exports

pub use gadget::GadgetStrategy;

pub use scalar::ScalarStrategy;

Modules

gadget
Strategy for zero-knowledge plonk circuits

scalar
Strategy for scalars

Traits

Strategy
Defines the Hades252 strategy algorithm.